Added pom.xml which will make it more easy to build the .mpk and updated the .jar files
