+++ src/EmailModuleWithTemplates/pom.xml
@@ -1,0 +1,181 @@
+<?xml version="1.0" encoding="UTF-8"?><project
+xmlns="http://maven.apache.org/POM/4.0.0"
+xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
+http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>com.mendix</groupId>
+  <artifactId>EmailModuleWithTemplates</artifactId>
+  <version>11.0.1</version>
+
+  <properties>
+    <mendixInstallDir>C:\Program Files\Mendix</mendixInstallDir>
+    <mendixVersion>8.12.7.15098</mendixVersion>
+  </properties>
+
+  <name>EmailModuleWithTemplates</name>
+  <url>https://docs.mendix.com/appstore/modules/email-with-templates</url>
+
+  <organization>
+    <name>Mendix</name>
+    <url>https://mendix.com</url>
+  </organization>
+
+  <repositories>
+    <repository>
+      <id>central</id>
+      <name>Maven Central</name>
+      <url>https://repo1.maven.org/maven2/</url>
+    </repository>
+  </repositories>
+
+  <build>
+    <outputDirectory>deployment/run/bin</outputDirectory>
+    <sourceDirectory>javasource</sourceDirectory>
+    <resources>
+      <resource>
+        <directory>javasource</directory>
+          <excludes>
+            <exclude>**/*.java</exclude>
+          </excludes>
+      </resource>
+    </resources>
+    <plugins>
+      <plugin>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.8.0</version>
+        <configuration>
+            <source>1.8</source>
+            <target>1.8</target>
+        </configuration>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-dependency-plugin</artifactId>
+        <executions>
+          <execution>
+            <id>copy-dependencies</id>
+            <phase>package</phase>
+            <goals>
+              <goal>copy-dependencies</goal>
+            </goals>
+            <configuration>
+              <outputDirectory>${project.basedir}/userlib</outputDirectory>
+              <overWriteReleases>false</overWriteReleases>
+              <overWriteSnapshots>false</overWriteSnapshots>
+              <overWriteIfNewer>true</overWriteIfNewer>
+              <excludeGroupIds>com.mendix</excludeGroupIds>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <artifactId>maven-clean-plugin</artifactId>
+        <version>3.1.0</version>
+        <configuration>
+          <filesets>
+            <fileset>
+              <directory>userlib</directory>
+            </fileset>
+          </filesets>
+        </configuration>
+      </plugin>
+      <plugin>
+        <artifactId>maven-antrun-plugin</artifactId>
+        <version>1.8</version>
+        <executions>
+          <execution>
+            <id>label-dependencies-libs</id>
+            <phase>install</phase>
+            <goals>
+              <goal>run</goal>
+            </goals>
+            <configuration>
+              <target>
+                <ant antfile="${project.basedir}/build.xml">
+                  <target name="createRequiredLibs"/>
+                </ant>
+              </target>
+            </configuration>
+          </execution>
+          <execution>
+            <id>export-module-package</id>
+            <phase>install</phase>
+            <goals>
+              <goal>run</goal>
+            </goals>
+            <configuration>
+              <target>
+                <ant antfile="${project.basedir}/build.xml">
+                    <target name="exportModulePackage"/>
+                  </ant>
+              </target>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+    </plugins>
+  </build>
+
+  <dependencies>
+    <dependency>
+     <groupId>ant-contrib</groupId>
+     <artifactId>ant-contrib</artifactId>
+     <version>1.0b3</version>
+    </dependency>
+    <dependency>
+     <groupId>com.mendix</groupId>
+     <artifactId>public-api</artifactId>
+     <version>${mendixVersion}</version>
+     <scope>system</scope>
+     <systemPath>${mendixInstallDir}\${mendixVersion}\runtime\bundles\com.mendix.public-api.jar</systemPath>
+    </dependency>
+    <dependency>
+     <groupId>com.mendix</groupId>
+     <artifactId>logging-api</artifactId>
+     <version>${mendixVersion}</version>
+     <scope>system</scope>
+     <systemPath>${mendixInstallDir}\${mendixVersion}\runtime\bundles\com.mendix.logging-api.jar</systemPath>
+    </dependency>
+    <dependency>
+     <groupId>javax.activation</groupId>
+     <artifactId>activation</artifactId>
+     <version>1.1.1</version>
+    </dependency>
+    <dependency>
+      <groupId>org.bouncycastle</groupId>
+      <artifactId>bcpg-jdk15on</artifactId>
+      <version>1.69</version>
+    </dependency>
+    <dependency>
+     <groupId>org.bouncycastle</groupId>
+     <artifactId>bcpkix-jdk15on</artifactId>
+     <version>1.69</version>
+    </dependency>
+    <dependency>
+     <groupId>org.bouncycastle</groupId>
+     <artifactId>bcprov-jdk15on</artifactId>
+     <version>1.69</version>
+    </dependency>
+    <dependency>
+      <groupId>commons-codec</groupId>
+      <artifactId>commons-codec</artifactId>
+      <version>1.15</version>
+    </dependency>
+    <dependency>
+     <groupId>org.apache.commons</groupId>
+     <artifactId>commons-email</artifactId>
+     <version>1.5</version>
+    </dependency>
+    <dependency>
+     <groupId>commons-io</groupId>
+     <artifactId>commons-io</artifactId>
+     <version>2.10.0</version>
+    </dependency>
+    <dependency>
+     <groupId>com.sun.mail</groupId>
+     <artifactId>javax.mail</artifactId>
+     <version>1.6.2</version>
+    </dependency>
+  </dependencies>
+</project>
